Add files via upload
DC1_new.py is the same as DC1.py, just changed slightly. I wanted to give it a new name so that we could keep the old DC1.py file incase it ends up being better. If you run main.py, you will see that it classifies the reviews and tells you the total number of misclassified reviews as well as the percentage of misclassfied reviews. Right now, the misclassified percentage is a little bit less than %50. Not sure if that is good or bad, might need to re-evaluate data collection